+++ services/web/server/src/simcore_service_webserver/version_control_handlers_base.py
@@ -1,6 +1,6 @@
 import logging
 from functools import wraps
-from typing import Any, Callable, Optional, Type
+from typing import Any, Type
 
 import orjson
 from aiohttp import web
@@ -8,7 +8,6 @@ from aiohttp.web_exceptions import HTTPException
 from pydantic.error_wrappers import ValidationError
 from pydantic.main import BaseModel
 from servicelib.aiohttp.typing_extension import Handler
-from yarl import URL
 
 from .projects.projects_exceptions import ProjectNotFoundError
 from .rest_utils import RESPONSE_MODEL_POLICY
@@ -74,29 +73,6 @@ def handle_request_errors(handler: Handler) -> Handler:
             ) from err
 
     return wrapped
-
-
-def create_url_for_function(request: web.Request) -> Callable:
-    app = request.app
-
-    def url_for(router_name: str, **params) -> Optional[str]:
-        try:
-            rel_url: URL = app.router[router_name].url_for(
-                **{k: f"{v}" for k, v in params.items()}
-            )
-            url = (
-                request.url.origin()
-                .with_scheme(
-                    request.headers.get("X-Forwarded-Proto", request.url.scheme)
-                )
-                .with_path(str(rel_url))
-            )
-            return f"{url}"
-
-        except KeyError:
-            return None
-
-    return url_for
 
 
 # FIXME: access rights using same approach as in access_layer.py in storage.
